+++ lib/msal-common/test/account/ClientInfo.spec.ts
@@ -48,6 +48,22 @@ describe("ClientInfo.ts Class Unit Tests", () => {
                             return input;
                     }
                 },
+                base64UrlEncode(input: string): string {
+                    switch (input) {
+                        case '{"kid": "XnsuAvttTPp0nn1K_YMLePLDbp7syCKhNHt7HjYHJYc"}':
+                            return "eyJraWQiOiAiWG5zdUF2dHRUUHAwbm4xS19ZTUxlUExEYnA3c3lDS2hOSHQ3SGpZSEpZYyJ9";
+                        default:
+                            return input;
+                    }
+                },
+                encodeKid(input: string): string {
+                    switch (input) {
+                        case "XnsuAvttTPp0nn1K_YMLePLDbp7syCKhNHt7HjYHJYc":
+                            return "eyJraWQiOiAiWG5zdUF2dHRUUHAwbm4xS19ZTUxlUExEYnA3c3lDS2hOSHQ3SGpZSEpZYyJ9";
+                        default:
+                            return input;
+                    }
+                },
                 async getPublicKeyThumbprint(): Promise<string> {
                     return TEST_POP_VALUES.KID;
                 },