+++ samples/msal-browser-samples/TypescriptTestApp2.0/utils/copyViews.js
@@ -1,3 +1,0 @@
-const shell = require("shelljs");
-
-shell.cp("-R", "src/views", "dist/");